// types/api.types.ts
export const EmailParticipantSchema = z.object({
  name: z.string().optional(),
  email: z.email(),
});
// ...
export const CreateDraftSchema = z.object({
  to: z.array(EmailParticipantSchema).optional(),
  cc: z.array(EmailParticipantSchema).optional(),
  bcc: z.array(EmailParticipantSchema).optional(),
  subject: z.string().optional(),
  body: z.string().optional(),
  bodyHtml: z.string().optional(),
  replyToEmailId: z.string().optional(),
});

export const UpdateDraftSchema = z.object({
  to: z.array(EmailParticipantSchema).optional(),
  cc: z.array(EmailParticipantSchema).optional(),
  bcc: z.array(EmailParticipantSchema).optional(),
  subject: z.string().optional(),
  body: z.string().optional(),
  bodyHtml: z.string().optional(),
});